pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beat_grid"
version = "0.1.0"
description = "Workflow 01 audio prep scaffold for beat grid and micro-timing output"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6,<3",
  "pyyaml>=6.0",
  "typer>=0.12",
  "jsonschema>=4.21",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[project.scripts]
audio-prep = "audio_prep.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]
